d_scripts_proj/d_scripts/requirements.txt
Django==4.0
djangorestframework==3.13.1
psycopg2-binary==2.9.3
python-dotenv==0.20.0
dj-rest-auth[with_jwt]==2.2.0
djangorestframework-simplejwt==5.1.0
gunicorn==20.1.0
django-cors-headers==3.13.0
